+++ mitra_models/migrations/schema.sql
@@ -226,6 +226,7 @@ CREATE TABLE notification (
     sender_id UUID NOT NULL REFERENCES actor_profile (id) ON DELETE CASCADE,
     recipient_id UUID NOT NULL REFERENCES user_account (id) ON DELETE CASCADE,
     post_id UUID REFERENCES post (id) ON DELETE CASCADE,
+    reaction_id UUID REFERENCES post_reaction (id) ON DELETE CASCADE,
     event_type SMALLINT NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now()
 );