+++ src/mastodon_api/notifications/types.rs
@@ -5,6 +5,7 @@ use mitra_models::notifications::types::{EventType, Notification};
 
 use crate::mastodon_api::{
     accounts::types::Account,
+    custom_emojis::types::CustomEmoji,
     pagination::PageSize,
     statuses::types::Status,
 };
@@ -20,6 +21,12 @@ pub struct NotificationQueryParams {
     pub limit: PageSize,
 }
 
+#[derive(Serialize)]
+pub struct EmojiReaction {
+    content: String,
+    emoji: Option<CustomEmoji>,
+}
+
 /// https://docs.joinmastodon.org/entities/notification/
 #[derive(Serialize)]
 pub struct ApiNotification {
@@ -28,10 +35,10 @@ pub struct ApiNotification {
     #[serde(rename = "type")]
     pub event_type: String,
 
-    pub created_at: DateTime<Utc>,
-
     pub account: Account,
     pub status: Option<Status>,
+    pub reaction: Option<EmojiReaction>,
+    pub created_at: DateTime<Utc>,
 }
 
 impl ApiNotification {
@@ -52,7 +59,8 @@ impl ApiNotification {
             EventType::Follow => "follow",
             EventType::FollowRequest => "follow_request",
             EventType::Reply => "reply",
-            EventType::Reaction => "favourite",
+            EventType::Reaction if notification.reaction_content.is_none() => "favourite",
+            EventType::Reaction => "emoji_reaction",
             EventType::Mention => "mention",
             EventType::Repost => "reblog",
             EventType::Subscription => "subscription",
@@ -61,12 +69,24 @@ impl ApiNotification {
             EventType::Move => "move",
             EventType::SignUp => "admin.sign_up",
         };
+        let maybe_reaction = if let Some(content) = notification.reaction_content {
+            let maybe_custom_emoji = notification.reaction_emoji
+                .map(|emoji| CustomEmoji::from_db(base_url, emoji));
+            let reaction = EmojiReaction {
+                content,
+                emoji: maybe_custom_emoji,
+            };
+            Some(reaction)
+        } else {
+            None
+        };
         Self {
             id: notification.id.to_string(),
             event_type: event_type_mastodon.to_string(),
-            created_at: notification.created_at,
             account,
             status,
+            reaction: maybe_reaction,
+            created_at: notification.created_at,
         }
     }
 }